+++ src/App.tsx
@@ -3,12 +3,12 @@ import './App.css';
 import MyNavbar from './components/MyNavbar';
 import './i18n/config';
 import { useEffect } from 'react';
-import Dashboard from './components/Dashboard';
+import Dashboard from './components/Dashboard/Dashboard';
 import { routes } from './Util/Global';
 import Footer from './components/Footer';
-import Accounts from './components/Accounts';
+import Accounts from './components/Account/Accounts';
 import Auth from './components/Auth';
-import Account from './components/AccountComponent';
+import Account from './components/Account/AccountComponent';
 import ConfirmMail from './components/ConfirmMail';
 import { supabase } from './supabaseClient';
 import { useSelector } from 'react-redux';
