+++ src/components/Account/NewsMessageModalContent.tsx
